Refactor core sprite logic so it can be reused on server
Without making an API call
